requirements.txt
pydantic==2.7.1
pandas==2.2.2
apache-airflow[celery]==2.9.1
